Added population size estimation methods for binomial model
The methods include 1) 'constant': The user can define a preknown fixed population size, 2) 'factor': The population size is taken as maximum observed y-value times a factor, 3) 'scale': The population size is estimated such that the observed values are distributed around a pre-defined success probability. Additionally, a bug for the normally distributed model was fixed.