Addendum for serialization
The stan_fit attribute of the backend has been necessary for the predict routine. In order to avoid serializing this complex object, the model_backend now includes a boolean attribute 'sample', which will be serialized and indicates wheter the model used the Laplace algorithm for sampling around the MLE solution. Also a demo work flow of the serialization was added to main.py