Added and modified available distribution models
*General Changes*
**Dispersion proxy**
All distributions that have a scale parameter except normal distribution controlling dispersion now internally use a dispersion proxy parameter kappa that scales dispersion with respect to a baseline model (e.g. Poisson is baseline for negative binomial or binomial is baseline for beta-binomial). With kappa = 0 the variance equals that of the baseline model. Placing a regularizing exponential prior on kappa keeps the dispersion from inflating. As a result
- the scale initial guess of all these models was removed
- Estimating the dispersion factor for the negative binomial model was removed

**Stan code**
All models that did not have GLM functions, had some variables created and expressions evaluated within a for loop in the ``model`` block. These expressions were vectorized and moved to the ``transformed parameters`` block.

**models.py**
- Added Fallback to Newton when BFGS fails. This prevents most RunTimeErrors
- Added compile=True flag to install_cmdstan. This automatically installs the C++ Toolchain necessary to compile CmdStan on Windows machines

*New models*
- Added beta and beta-binomial distribution
- for beta-distribution: Stan files uses beta_proportional() which is parametrized using the expectation value
- logit chosen as link function for both models
- Added models to models.py
- Added testscripts for beta distribution